requirements.txt
absl-py==0.11.0
astunparse==1.6.3
backcall==0.2.0
blis==0.7.4
cachetools==4.2.0
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
cupy==8.3.0
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
fastrlock==0.5
flatbuffers==1.12
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-2.3.0/fr_core_news_sm-2.3.0.tar.gz
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.18.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
matplotlib==3.3.3
murmurhash==1.0.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pkg-resources==0.0.0
plac==1.1.3
preshed==3.0.5
prompt-toolkit==3.0.10
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
spacy==2.3.5
spacy-lookups-data==0.3.2
srsly==1.0.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
thinc==7.4.5
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.2
wasabi==0.8.1
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.12.1
